Service/pdf-service/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
python-multipart==0.0.22
pillow==12.1.1
pytesseract==0.3.13
pymupdf==1.27.2